[helpers] consolidate and move around
