+++ aws_s3_rules/aws_s3_access_error.py
@@ -1,4 +1,4 @@
-from panther_oss_helpers import pattern_match
+from panther_base_helpers import pattern_match
 
 # https://docs.aws.amazon.com/AmazonS3/latest/API/ErrorResponses.html
 HTTP_STATUS_CODES_TO_MONITOR = {
