+++ aws_s3_rules/aws_s3_insecure_access.py
@@ -1,4 +1,4 @@
-from panther_oss_helpers import pattern_match
+from panther_base_helpers import pattern_match
 
 
 def rule(event):
